+++ goleveldb_batch.go
@@ -19,22 +19,22 @@ func newGoLevelDBBatch(db *GoLevelDB) *goLevelDBBatch {
 	}
 }
 
-func (b *goLevelDBBatch) assertOpen() {
-	if b.batch == nil {
-		panic("batch has been written or closed")
-	}
-}
-
 // Set implements Batch.
-func (b *goLevelDBBatch) Set(key, value []byte) {
-	b.assertOpen()
+func (b *goLevelDBBatch) Set(key, value []byte) error {
+	if b.batch == nil {
+		return errBatchClosed
+	}
 	b.batch.Put(key, value)
+	return nil
 }
 
 // Delete implements Batch.
-func (b *goLevelDBBatch) Delete(key []byte) {
-	b.assertOpen()
+func (b *goLevelDBBatch) Delete(key []byte) error {
+	if b.batch == nil {
+		return errBatchClosed
+	}
 	b.batch.Delete(key)
+	return nil
 }
 
 // Write implements Batch.
@@ -48,20 +48,22 @@ func (b *goLevelDBBatch) WriteSync() error {
 }
 
 func (b *goLevelDBBatch) write(sync bool) error {
-	b.assertOpen()
+	if b.batch == nil {
+		return errBatchClosed
+	}
 	err := b.db.db.Write(b.batch, &opt.WriteOptions{Sync: sync})
 	if err != nil {
 		return err
 	}
 	// Make sure batch cannot be used afterwards. Callers should still call Close(), for errors.
-	b.Close()
-	return nil
+	return b.Close()
 }
 
 // Close implements Batch.
-func (b *goLevelDBBatch) Close() {
+func (b *goLevelDBBatch) Close() error {
 	if b.batch != nil {
 		b.batch.Reset()
 		b.batch = nil
 	}
+	return nil
 }